chore: Removed method from protocol
